Fix elapsed time format
